requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.3
django-filter==24.1
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exrex==0.11.0
filelock==3.13.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==24.0
pandas==2.2.1
platformdirs==4.2.0
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-futures==1.0.1
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stem==1.8.2
torrequest==0.1.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0

djangorestframework~=3.15.1